Adding Colors, Fonts, Padding, Margins and Pseudo Classes
After finishing with the background image of the webpage, started setting up the colors for the nav and text. After aligning the text; using padding and margins to make the page look more professional and aligning the text to how it should be. The Pseudo classes were used on the nav to change the color and font size as directed (font-size will be in next commit).